+++ IHM/lib/Elements/src/RelayElement.h
@@ -42,14 +42,14 @@ class RelayElement : public Element {
             char buf[PWM_LABEL_BUFFER_SIZE];
             formatShortStatusLabel(buf, config.state);
             status.setLabel(buf);
-            led.setState(config.state ? 1 : 0);
+            led.setState(config.state ? LED_ON : LED_OFF);
         }
 
     public:
         RelayElement(uint16_t x, uint16_t y, uint16_t w, uint16_t h, Display *tft, Relay *relay, uint8_t relayIndex):
-                                            Element(PWM_SELECTION, tft),
-                                            led(x + RELAYELEMENT_LED_X, y + h / 2, RELAYELEMENT_LED_TEXTSIZE, RELAYELEMENT_LED_RADIUS, 0, "Rele", 0, tft),
-                                            status(x + RELAYELEMENT_STATUS_X, y + h / 2 - 8, RELAYELEMENT_STATUS_TEXTSIZE, h, 1, "Desl.", 1, tft),
+                                            Element(tft),
+                                            led(x + RELAYELEMENT_LED_X, y + h / 2, RELAYELEMENT_LED_TEXTSIZE, RELAYELEMENT_LED_RADIUS, LED_OFF, "Rele", 0, tft),
+                                            status(x + RELAYELEMENT_STATUS_X, y + h / 2 - 8, RELAYELEMENT_STATUS_TEXTSIZE, h, LABEL_ON, "Desl.", 1, tft),
                                             relay(relay), relayIndex(relayIndex)
         {
             Element::x = x;
@@ -61,7 +61,7 @@ class RelayElement : public Element {
             led.setLocation((int)right);
         }
 
-        Element update() {
+        void update() {
             if (!isShown) {
                 isShown = true;
                 if (isSelected) {
@@ -70,17 +70,16 @@ class RelayElement : public Element {
                     tft->fillRect(x, y, RELAYELEMENT_WIDTH, h, BLACK);
                 }
                 tft->drawFastHLine(x, y + h, RELAYELEMENT_WIDTH, TFT_ORANGE);
-                led.setState(config.state ? 1 : 0);
+                led.setState(config.state ? LED_ON : LED_OFF);
             }
             led.update();
             status.update();
-            return *this;
         }
 
         void show() {
             isShown = false;
             refreshDisplay();
-            status.setState(1);
+            status.setState(LABEL_ON);
         }
 
         void setSelected(bool isSelected) {